Update sgs.py
Fixed bug of setting the mean kriging value in SGS. Instead now the mean value and standard deviation from kriging is used to draw a random sample from, which is then set as the SGS value. Plus experimenting with passing a distance measure to select a circular neighbourhood to limit influence of points lying far away.